Simplify ATen Build (#3496)
* THS build change

* merge THCS into ATen build

* THCUNN build change over

* update THNN build

* move THC build to ATen, as well as some of the accumulated top level config from other TH* libraries

* TH library build merged into ATen, and warnings fixes.

* fix magma support checking

* check cuda early

* fall back to GCC atomics if C11 atomics have issues.

* fix install name

* disable openmp in files that also include stdatomic.h

* make sure LAPACK is visible to TH build file.

+++ aten/src/THC/generic/THCTensorMath.cu
@@ -224,7 +224,7 @@ void THCTensor_(catArray)(THCState *state, THCTensor *result,
         stackInputs[j].offset = offset;
         stackInputs[j].dimSize = dimSize;
         stackInputs[j].nElements = THCTensor_(nElement)(state, inputs[i+j]);
-        cohortMax = cohortMax > stackInputs[j].nElements ? cohortMax : stackInputs[j].nElements;
+        cohortMax = cohortMax > (int) stackInputs[j].nElements ? cohortMax : (int) stackInputs[j].nElements;
 
         // update offset
         offset += dimSize;
